Complete rewrite of algorithm
The previous algorithm was a confusing mess and wasn't really all that
efficient anyway, because it would just iterate regexes any time they
started with a wildcard, even if the ends could be anchored.

The new algorithm uses a trie compiled into a regex, using perl's
regex code-execution feature to identify which of the patterns
got matched.  It also has an optimization to reverse the subject
and match backward form the end in the case that the pattern
begins with a wildcard but ends with anchored text.

The constant factor of the algorithm might be higher, but the
big-O should be better for a lot of common cases.