Implement better globstar semantics, like /**/ matching /
Globstar from the shell treats '/' characters as separators instead
of like characters to be matched.  /** at the end can match an
empty string, and /**/ in the middle can match a single slash,
and so on.

I also got rid of the 'matchers' in favor of plain regex-refs.

Added more tests and fixed various bugs.