New ground-up implementation, again
This time, the official behavior is to check the patterns in the
order they are given, unless the user asks to sort them into the
optimal longest-prefix order.

Then, the correct behavior is easy to test by the naieve algorithm
of converting glob to regex and iterating them.  The advanced
algorithm will look for opportunities to do prefix-trie matching
on rules for any case that it doesn't affect the evaluation order.

Also, I started into the set of features for controlling which
shell syntax is enabled and letting users override the characters.